day_04/practicum/simulverse-main/requirements.txt
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.0
bson==0.5.10
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
ecdsa==0.18.0
fastapi==0.82.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==3.0.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
uvicorn==0.18.3
